Fixes float64 -> object bug
Calculations with large integers (e.g. 10 ** 20 instead of 1e20) result in NumPy casting to object dtype instead of float64. This triggers errors when performing further calculations or checks, like isfinite(). When fitting such data, the fit failed with an exception. This change casts the results back to float64.

+++ src/tailor/data_model.py
@@ -299,7 +299,7 @@ class DataModel(QtCore.QAbstractTableModel):
                     f"ERROR: {exc}: {msg}.", msecs=MSG_TIMEOUT
                 )
         elif output is not None:
-            self._data[col_name] = output
+            self._data[col_name] = output.astype("float64")
             return True
         else:
             print("No evaluation error but no output.")
